refactor(state): status -> state package name 변경
