+++ src/main/java/blackjack/domain/state/Blackjack.java
@@ -1,4 +1,4 @@
-package blackjack.domain.status;
+package blackjack.domain.state;
 
 public class Blackjack extends Finished {
     public Blackjack(Cards cards) {
