+++ src/main/java/blackjack/domain/state/Bust.java
@@ -1,4 +1,4 @@
-package blackjack.domain.status;
+package blackjack.domain.state;
 
 public class Bust extends Finished {
     public Bust(Cards cards) {
