+++ src/main/java/blackjack/domain/state/StartRound.java
@@ -1,4 +1,4 @@
-package blackjack.domain.status;
+package blackjack.domain.state;
 
 import blackjack.domain.card.Card;
