+++ src/main/java/blackjack/domain/state/Stay.java
@@ -1,4 +1,4 @@
-package blackjack.domain.status;
+package blackjack.domain.state;
 
 public class Stay extends Finished{
     public Stay(Cards cards) {
