+++ src/test/java/blackjack/domain/state/StartRoundTest.java
@@ -1,4 +1,4 @@
-package blackjack.domain.status;
+package blackjack.domain.state;
 
 import org.junit.jupiter.api.Test;
